+++ src/gource.h
@@ -89,7 +89,7 @@ class Gource : public SDLApp {
     RDirNode* root;
 
     std::string displaydate;
-    int date_x_offset;
+    float date_x_offset;
 
     FXFont font, fontlarge, fontmedium;
